Add control and move transformer instances to companion
