+++ core/src/main/scala/cats/data/IndexedStateT.scala
@@ -246,23 +246,6 @@ private[data] sealed abstract class IndexedStateTInstances extends IndexedStateT
   implicit def catsDataAlternativeForIndexedStateT[F[_], S](implicit FM: Monad[F],
     FA: Alternative[F]): Alternative[IndexedStateT[F, S, S, ?]] with Monad[IndexedStateT[F, S, S, ?]] =
     new IndexedStateTAlternative[F, S] { implicit def F = FM; implicit def G = FA }
-
-  implicit def catsErrorControlForStateT[F[_], G[_], S, E]
-  (implicit E: ErrorControl[F, G, E], M: Monad[G]): ErrorControl[StateT[F, S, ?], StateT[G, S, ?], E] =
-    new ErrorControl[StateT[F, S, ?], StateT[G, S, ?], E] {
-      implicit val F: MonadError[F, E] = E.monadErrorF
-
-      val monadErrorF: MonadError[StateT[F, S, ?], E] = IndexedStateT.catsDataMonadErrorForIndexedStateT
-      val monadG: Monad[StateT[G, S, ?]] = IndexedStateT.catsDataMonadForIndexedStateT(M)
-
-      def accept[A](ga: StateT[G, S, A]): StateT[F, S, A] = ga.mapK(new (G ~> F) {
-        def apply[T](ga: G[T]): F[T] = E.accept(ga)
-      })
-
-      def controlError[A](fa: StateT[F, S, A])(f: E => StateT[G, S, A]): StateT[G, S, A] =
-        IndexedStateT(s => E.controlError(fa.run(s))(e => f(e).run(s)))
-
-    }
 }
 
 private[data] sealed abstract class IndexedStateTInstances1 extends IndexedStateTInstances2 {
